Fixed CategoriesController getById method to check if equal to null throw an exception with statuc code (not found)
Fixed CategoryService getById to find the Id or else return NULL

+++ src/main/java/org/yearup/controllers/CategoriesController.java
@@ -18,13 +18,13 @@ import java.util.List;
 @RestController
 // add the annotation to make this controller the endpoint for the following url
     // http://localhost:8080/categories
-@RequestMapping("/categories")
+@RequestMapping("/api/categories")
 // add annotation to allow cross site origin requests
 @CrossOrigin
 public class CategoriesController {
+
     private CategoryService categoryService;
     private ProductService productService;
-
 
     // create an Autowired constructor to inject the categoryService and productService
     @Autowired
@@ -35,6 +35,7 @@ public class CategoriesController {
 
     // add the appropriate annotation for a get action
     @GetMapping
+    @PreAuthorize("permitAll()")
     public List<Category> getAll()
     {
         // find and return all categories
@@ -43,6 +44,7 @@ public class CategoriesController {
 
     // add the appropriate annotation for a get action
     @GetMapping("/{id}")
+    @PreAuthorize("permitAll()")
     public Category getById(@PathVariable int id)
     {
         // get the category by id
@@ -59,7 +61,11 @@ public class CategoriesController {
     public List<Product> getProductsById(@PathVariable int categoryId)
     {
         // get a list of product by categoryId
-        return productService.listByCategoryId(categoryId);
+        List<Product> productById = productService.listByCategoryId(categoryId);
+        if (productById == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, " Product not found");
+        }
+        return productById;
     }
 
     // add annotation to call this method for a POST action
@@ -70,7 +76,7 @@ public class CategoriesController {
     {
         // insert the category and return it with status 201 Created
         Category addedCategory = categoryService.create(category);
-        return new ResponseEntity<>(addedCategory, HttpStatus.CREATED);
+        return ResponseEntity.status(HttpStatus.CREATED).body(addedCategory);
     }
 
     // add annotation to call this method for a PUT (update) action - the url path must include the categoryId
@@ -81,9 +87,11 @@ public class CategoriesController {
     {
         // update the category by id and return the updated category (200 OK)
         Category updatedCategory = categoryService.update(id, category);
-        return new ResponseEntity<>(updatedCategory, HttpStatus.OK);
+        if (updatedCategory == null){
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(updatedCategory);
     }
-
 
     // add annotation to call this method for a DELETE action - the url path must include the categoryId
     // add annotation to ensure that only an ADMIN can call this function
@@ -92,7 +100,10 @@ public class CategoriesController {
     public ResponseEntity<Void> deleteCategory(@PathVariable int id)
     {
         // delete the category by id and return status 204 No Content
+        if (categoryService.getById(id) == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
         categoryService.delete(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return ResponseEntity.noContent().build();
     }
 }
